update: gen_mtメソッドを別モジュールに切り出し
